% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{Performs a oneway analysis}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{???}

\item{data}{???}
}
\value{
a object of class "oneway"
}
\description{
This function performs a oneway analysis of variance and returns a object of
class "oneway"
}
\examples{
oneway(formula, dataset) ????
}
