% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnewayR.R
\docType{package}
\name{OnewayR}
\alias{OnewayR}
\title{OnewayR}
\description{
This is a package for performing
a oneway analysis of variance,
creating an object with a class of "oneway" with additional
print, plot, and summary functions.
}
