% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plots the results}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{object class of type oneway}

\item{...}{additional parameters that are being passed to the print function}

\item{color}{input a color of choice using col = "x". If not specified the
default color is skyblue.}
}
\value{
boxplot plotting the results of the analysis oneway.
}
\description{
This function plots the results of the analysis oneway
}
\examples{
plot.oneway(oneway, ...)
plot.oneway(oneway, col = "red", ...)
}
