% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Prints results}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{a numeric vector}
}
\value{
a numeric vector
}
\description{
This function prints the result of the analysis
}
\examples{
add(2,3)
add(mtcars$mpg, mtcars$hp)
}
