% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Prints a summary}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{object class of type oneway}

\item{...}{additional parameters that are being passed to the print function.}
}
\value{
a data frame that summarizes the analysis of oneway.
}
\description{
This function prints a summary of the analysis oneway.
}
\examples{
plot.oneway(oneway, ...)
}
