% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{One-Way ANOVA Test}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" (i.e, y ~ model).
See \code{formula} for more details.}

\item{data}{data is presented as data.frame}
}
\value{
an object of class "oneway"
}
\description{
This oneway function performs a one-way analysis of variance
for two or more numeric variables.
}
\details{
This function uses \code{lm} to describe the linear relationship and \code{aggregate}
to obtain \code{length}, \code{mean}, and \code{sd}.
}
\examples{
\dontrun{

#perform one-way ANOVA

mydataframe <- oneway(mpg ~ wt, mtcars)

#perform one-way ANOVA with more than two groups

mydataframe <- oneway(mpg ~ wt + hp + disp + qsec)
}

}
