% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{One-Way ANOVA Test}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" i.e, y ~ model.
Use "+" symbol to denote more than one numeric variable.}

\item{data}{data is presented as data.frame}
}
\value{
an object of class "oneway"
}
\description{
This oneway function performs a one-way analysis of variance
for two or more numeric variables.
}
\details{
This function uses both \code{lm} and \code{aggregate} to fit linear models
and obtain \code{length}, \code{mean}, and \code{sd}, respectively.
}
\examples{
\dontrun{

#perform one-way ANOVA

mydataframe <- oneway(mpg ~ wt, mtcars)
}

}
