% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plots results}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{???? If not class "oneway" the function
will stop and state that class must be oneway.}

\item{...}{parameters passed to the oneway function.}

\item{color}{input a color of choice using col = "x". If not specified the
default color is skyblue.}
}
\value{
a boxplot ???
}
\description{
This function prints the results of the analysis
}
\examples{
plot.oneway(x, col = "" , ...) ???
}
