% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plotting One-Way ANOVA Results}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{an object of class oneway}

\item{col}{color of each box}

\item{...}{optional arguments: See \code{boxplot} for more details.}
}
\description{
Produces box plots of the One-Way ANOVA results
}
\examples{
\dontrun{

plot(mydataframe, col="blue")

}


}
