% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plotting One-Way ANOVA}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{an object of class oneway}

\item{col}{color of each box}

\item{...}{optional arguments: See \code{plot}.}
}
\value{
boxplot of One-Way ANOVA
}
\description{
Boxplots the One-Way ANOVA results
}
\details{
The \code{lm} and \code{aggregate} is used to fit linear models
and obtain \code{length}, \code{mean}, and \code{sd}, respectively.
}
\examples{
\dontrun{

plot(mydataframe, col="blue")

}


}
