% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot Oneway Analysis}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments (not currently used)}
}
\value{
a boxplot
}
\description{
the \code{plot.oneway} function plots the results of the oneway analysis
}
\examples{
x <- oneway(mpg~cyl, mtcars)
plot(x)

}
\seealso{
\link{oneway},
\link{print.oneway},
\link{summary.oneway}
}
