% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Prints the results}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{object class of type oneway}

\item{...}{additional parameters that are being passed to the print function}
}
\value{
the results of the analysis oneway printed in the console
}
\description{
This function prints the result of the analysis oneway
}
\examples{
print.oneway(oneway,...)
}
