% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Oneway Printing}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{object of class oneway}

\item{...}{optional arguments to \code{print} methods}
}
\description{
Prints one-way ANOVA results.
}
\details{
Details the outcome of the one-way ANOVA.
}
