% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print Oneway Analysis}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments (not currently used)}
}
\value{
summary statistics and anova results
}
\description{
The \code{print.oneway} function prints the results of the oneway analysis:
the summary statistics (count, mean, and standard deviations of each group) and
the anova results (analysis of variance)
}
\examples{
x <- oneway(mpg~cyl, mtcars)
print(x)

}
\seealso{
\link{oneway},
\link{plot.oneway},
\link{summary.oneway}
}
