% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Plots the results}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{?? If not class "oneway" the function
will stop and state that class must be oneway.}

\item{...}{parameters passed to the oneway function.}
}
\value{
a boxplot ?????
}
\description{
This function prints a summary of the analysis
}
\examples{
plot.oneway(x, col = "" , ...) ???
}
