% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize One-Way Anova}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class oneway}

\item{...}{optional arguments for the function: See \code{summary}.}
}
\value{
summary of one-way ANOVA
}
\description{
This summary.oneway function takes object of class oneway and prints out
the results of the statistical test.
}
\examples{
\dontrun{

summary.oneway(mydataframe)

}


}
