% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize Oneway Analysis}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{additional arguments (not currently used)}
}
\value{
anova results
}
\description{
The \code{summary.oneway} function prints a summary of the analysis of variance
}
\examples{
x <- oneway(mpg~cyl, mtcars)
summary.oneway(x)

summary(x)

}
\seealso{
\link{oneway},
\link{print.oneway},
\link{plot.oneway}
}
